backend/requirements.txt
fastapi
pytesseract
uvicorn
google-genai
pydantic>=2.0
python-multipart
pillow
python-dotenv